Split out code from main runner
