+++ test/test_nyp_tracks.py
@@ -1,7 +1,8 @@
-import nypTracks
 from lib.parse_utils import Departure 
+from lib.tracks import changes_as_events, pipeline
 
 import unittest
+from unittest.mock import MagicMock, patch
 
 TRENTON_123=Departure("11:00", "Trenton", "", "NEC", "123")
 BAY_HEAD_345=Departure("11:10", "Bay Head", "", "whatever", "345")
@@ -10,7 +11,7 @@ class TestNypTracks(unittest.TestCase):
     def test_new_departures(self):
         old = [TRENTON_123]
         new = [TRENTON_123, BAY_HEAD_345]
-        events = nypTracks.changes_as_events(old, new)
+        events = changes_as_events(old, new)
         self.assertEqual(len(events), 1)
         self.assertEqual(events[0].action, "added")
         self.assertEqual(events[0].context, "_window")
@@ -19,7 +20,7 @@ class TestNypTracks(unittest.TestCase):
     def test_dropped_departures(self):
         old = [TRENTON_123, BAY_HEAD_345]
         new = [BAY_HEAD_345]
-        events = nypTracks.changes_as_events(old, new)
+        events = changes_as_events(old, new)
         self.assertEqual(len(events), 1)
         self.assertEqual(events[0].action, "dropped")
         self.assertEqual(events[0].context, "_window")
@@ -30,12 +31,18 @@ class TestNypTracks(unittest.TestCase):
         assignment = Departure(**TRENTON_123.__dict__)
         assignment.track = "1"
         new = [assignment, BAY_HEAD_345]
-        events = nypTracks.changes_as_events(old, new)
+        events = changes_as_events(old, new)
         self.assertEqual(len(events), 1)
         self.assertEqual(events[0].action, "changed")
         self.assertEqual(events[0].context, "track")
         self.assertEqual(events[0].new.track, "1")
         self.assertEqual(events[0].old.track, "")
+    
+    @patch('urllib.request.urlopen', MagicMock(side_effect=Exception('mock triggered')))
+    def test_pipeline(self):
+        pipeline()
+
 
 if __name__ == '__main__':
     unittest.main()
+
